Pass location param to mapped tileID method
Upstream change to support alternate location data for benchmarks to run
against.
