+++ bench/styles/benchmarks.js
@@ -1,7 +1,7 @@
 import mapboxgl from '../../src';
 import accessToken from '../lib/access_token';
-import locations from '../lib/style_locations';
-
+import locationsWithTileID from '../lib/locations_with_tile_id';
+import styleBenchmarkLocations from '@mapbox/gazetteer/mapbox-streets/style-benchmark-locations.json';
 import StyleLayerCreate from '../benchmarks/style_layer_create';
 import Validate from '../benchmarks/style_validate';
 import Layout from '../benchmarks/layout';
@@ -10,6 +10,8 @@ import QueryPoint from '../benchmarks/query_point';
 import QueryBox from '../benchmarks/query_box';
 
 import getWorkerPool from '../../src/util/global_worker_pool';
+
+const locations = locationsWithTileID(styleBenchmarkLocations);
 
 mapboxgl.accessToken = accessToken;
